Organizei os arquivos de CadastroUsuario e Auxilair, facilitando o uso das funções, e retirando as funções específicas de cada um deles do main
