+++ src/dados/usuarios.txt
@@ -1,2 +1,5 @@
 RAPHAEL,123564!,Nubank 200
 NICOLASMNL,kjashd*,BB 1200,Inter 1400
+JOAOTEIMOSO,abcd123*
+MARIADASGRACAS,123abcd!
+JOAOSILVA,hrvn1234@#,Inter 800
