+++ src/main.hs
@@ -1,6 +1,7 @@
 import Auxiliar as Auxiliar
 import Usuario as Usuario
 import Banco as Banco
+import CadastroUsuario as CadastroUsuario
 import System.IO
 import System.Directory
 import Data.Char
@@ -36,13 +37,13 @@ menuLogin = do
 
 			arquivo <- readFile "dados/usuarios.txt"
 			
-			if existeLogin login arquivo
+			if Auxiliar.existeLogin login arquivo
 				then 
-					if verificaSenha login senha arquivo
+					if Auxiliar.verificaSenha login senha arquivo
 						then 
 							do
 							putStrLn ("Bem vindo " ++ login ++ "!\n")
-							menuUsuario (toUpperCase login) 
+							menuUsuario (Auxiliar.toUpperCase login) 
 						else 
 							do
 							putStrLn "Senha incorreta. Tente novamente\n"
@@ -70,7 +71,7 @@ menuLogin = do
 					do
 					putStrLn "Login ja cadastrado. Tente novamente com outro login\n"
 					menuLogin
-				else cadastraUsuario login senha
+				else CadastroUsuario.cadastraUsuario login senha
 			
 			menuLogin
 			
@@ -163,86 +164,4 @@ menuUsuario login = do
 			putStrLn "Opcao invalida!\n"
 			menuUsuario login
 
-
-
-----------------------------------------------------------------
---CadastroUsuario
-
--- Função que irá cadastrar o usuário num arquivo .txt, mantendo um registro dos usuarios cadastrados
-cadastraUsuario :: String -> String -> IO ()
-cadastraUsuario login senha = 
-    if validaCadastro login senha
-        then
-            do
-            appendFile "dados/usuarios.txt" ((toUpperCase login) ++ "," ++ senha ++ "\n")
-            putStrLn "Cadastro realizado com sucesso\n"
-			
-        else 
-			do
-			putStrLn "Nao foi possivel realizar o cadastro. Tente novamente\n"
-			
-
--- -- Função que vai ser responsável por validar o cadastro do Usuário
--- -- Login não pode ser vazio
--- -- Senha não pode ser vazia
--- -- senha possui 6 ou mais caracteres
-validaCadastro :: String -> String -> Bool
-validaCadastro "" senha = False
-validaCadastro login "" = False
-validaCadastro login senha = 
-    if length senha <= 6
-        then False
-        else not (null (filter (`elem` ['*', '!', '@', '/', '#']) senha))
-
-
-
-
-
--- carregaUsuarios :: [Usuario.Usuario]
--- carregaUsuarios = Auxiliar.getListaDeUsuarios 
-
-
-
-
-
-
 -------------------------------------------------------------------------------------
-
-
---Auxiliar
-
-
-
---Retorna a lista de um usuário cadastrado no sistema
---Caso o usuário não esteja cadastrado ele retorna uma lista vazia
-getListaLogin :: [[String]] -> String -> [[String]]
-getListaLogin lista login = Data.List.filter (login `elem`) lista
-
-existeLogin :: String -> String -> Bool
-existeLogin login arquivo = do
-    let lista_usuarios = ((Data.List.map ( splitOn ",") (lines arquivo)))
-    let lista_login = getListaLogin lista_usuarios (toUpperCase login)
-    if null lista_login
-        then False
-        else True
-
--- verifica se a senha de um determinado login esta correta ou nao
-verificaSenha :: String -> String -> String -> Bool 
-verificaSenha login senha arquivo = do
-	let lista_usuarios = ((Data.List.map ( splitOn ",") (lines arquivo)))
-	let lista_login = getListaLogin lista_usuarios (toUpperCase login)
-	if (lista_login !! 0) !! 1 == senha
-		then True 
-		else False
-
-
-exibirUsuario :: String -> String -> [String]
-exibirUsuario login arquivo = do
-	let lista_login = getListaLogin ((Data.List.map ( splitOn ",") (lines arquivo))) login
-	head lista_login
-
-
-
--- Recebe uma string e tranforma tudas as letras minusculas em maiusculas
-toUpperCase :: String -> String
-toUpperCase palavra = [toUpper x | x <- palavra]
